+++ lib/cheftacular/error.rb
@@ -5,12 +5,15 @@ class Cheftacular
       @options, @config  = options, config
     end
 
-    def is_valid_node_name_option?
+    def is_valid_node_name_option? strict_env=false
       raise "Too few arguments, please supply a node name" if @options['node_name'].nil?
 
       nodes = @config['getter'].get_true_node_objects(true)
 
-      nodes = @config['parser'].exclude_nodes( nodes, [{ if: { not_node: @options["node_name"] } }], true )
+      exclusion_args  = [{ if: { not_node: @options["node_name"] } }]
+      exclusion_args << { if: { not_env: @options['env'] } } if strict_env
+
+      nodes = @config['parser'].exclude_nodes( nodes, exclusion_args, true )
 
       raise "Node not found for #{ @options['node_name'] }" if nodes.empty?